Fix #44 (the flatten array bug)
The original bug was that if you did this

    require 'rr'
    include RR::Adapters::RRMethods
    mock(o = Object.new).some_random_method
    [o].flatten

then RR would complain with the following:

    NoMethodError: undefined method `to_ary' for #<Object:0x007feeba8336b8>

When you mock or stub a method on an object using RR, it will create a
method_missing method in the object's singleton class (assuming it is not there
already). This issue involves Array#flatten, which, in addition to calling
\#to_ary on each object in the array, also calls #method_missing on each object.
Now usually Array#flatten will swallow any NoMethodErrors that occur while
calling both #to_ary and #method_missing, but for some reason that is not
happening in this case. Regardless, we can combat this by supplying our own
version of #to_ary that returns nil. This tells Ruby to ignore our #to_ary and
use the default behavior in arrayifying our object in question.